refactor: :recycle: update components name
