+++ components/ContentDetails.tsx
@@ -1,10 +1,10 @@
 import { Icon } from '@/components';
 
 interface Props {
-  items: ArticleDetailsProps[];
+  items: ContentDetailsProps[];
 }
 
-function ArticleDetails({ items }: Props) {
+function ContentDetails({ items }: Props) {
   return (
     <div className="border-border-light flex w-full flex-wrap gap-x-7 gap-y-2 border px-6 py-2 md:w-fit lg:py-6">
       {items.map(item => (
@@ -24,4 +24,4 @@ function ArticleDetails({ items }: Props) {
   );
 }
 
-export default ArticleDetails;
+export default ContentDetails;
