+++ components/index.ts
@@ -2,4 +2,4 @@ export { default as Brand } from './Brand';
 export { default as Button } from './Button';
 export { default as Icon } from './Icon';
 export { default as Card } from './Card';
-export { default as ArticleDetails } from './ArticleDetails';
+export { default as ContentDetails } from './ContentDetails';
