+++ types/ContentDetailsProps.ts
@@ -1,4 +1,4 @@
-interface ArticleDetailsProps {
+interface ContentDetailsProps {
   icon: IconLabelProps;
   label: string;
 }
